refactor: 금액 범위 수정 (#114)
* refactor: 상품 수정 요청 필드 수정

상품 수정 요청 필드 검증 조건 수정

CHZZ-176

* fix: 금액 최소 및 최대범위 수정

모든 금액 관련 요청을 1000원 이상 200만원 이하로 수정

CHZZ-176

* refactor: 최소범위 제거 및 코드 스타일 수정

최소 조건 제거 및 코드 스타일 조정

CHZZ-176

* refactor: 결제 조건 수정

최소 결제 금액 조건을 1000의 배수 조건으로 수정

CHZZ-176

* refactor: 오타 수정

오타 수정

CHZZ-176

* style: 간격 수정

간격 수정

CHZZ-176

* fix: getter 추가

오타로 제거된 getter 추가

CHZZ-176
