+++ src/main/java/org/chzz/market/domain/product/dto/UpdateProductRequest.java
@@ -4,7 +4,7 @@ import static org.chzz.market.domain.auction.dto.request.BaseRegisterRequest.DES
 import static org.chzz.market.domain.product.entity.Product.Category;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import java.util.HashMap;
@@ -31,7 +31,7 @@ public class UpdateProductRequest {
     private Category category;
 
     @ThousandMultiple
-    @Min(value = 1000, message = "시작 가격은 최소 1,000원 이상, 1000의 배수이어야 합니다")
+    @Max(value = 2_000_000, message = "최소금액은 200만원을 넘을 수 없습니다")
     private Integer minPrice;
 
     @Builder.Default
